large_model_eval/requirements.txt
# 大语言模型评估系统依赖

# Web框架
fastapi
uvicorn[standard]==0.24.0
pydantic==2.5.0

# OpenAI API
openai

# 数据处理
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# 评估指标
rouge-score==0.1.2
bert-score==0.3.13
sacrebleu==2.3.1

# 文本处理
nltk==3.8.1
huggingface-hub==0.21.4
transformers==4.39.0
torch==2.1.0

# 配置文件
pyyaml==6.0.1

# 日志和监控
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 类型检查
mypy==1.7.0
